move base class out of Formula/ so it can't be installed
