+++ Formula/timescaledb@2.6.0.rb
@@ -1,4 +1,4 @@
-require_relative 'timescaledb_base'
+require_relative "../timescaledb_base"
 
 class TimescaledbAT260 < TimescaledbBase
   sha256 "058991ae9c312ed3e08356262103cfbe9d6380a94bb8a207710c31c38f2a2d5f"
