+++ Formula/timescaledb@2.6.1.rb
@@ -1,4 +1,4 @@
-require_relative 'timescale_base'
+require_relative "../timescaledb_base"
 
 class TimescaledbAT261 < TimescaledbBase
   sha256 "226af207105cc9e2d1744b303e475204a947047103f0e5697de4e8949e4dde75"
