+++ Formula/timescaledb@2.7.0.rb
@@ -1,4 +1,4 @@
-require_relative 'timescaledb_base'
+require_relative "../timescaledb_base"
 
 class TimescaledbAT270 < TimescaledbBase
   sha256 "9b30952da1186733f144c1ec3aa0c6332379207e5035e81037bd34e1ec9cefb3"
