+++ Formula/timescaledb@2.7.1.rb
@@ -1,4 +1,4 @@
-require_relative 'timescale_base'
+require_relative "../timescaledb_base"
 
 class TimescaledbAT271 < TimescaledbBase
   sha256 "595f7403bd8ab6b44742fcacf991a0ea26a43a4936dea1d5e7627e53de920fe0"
