+++ Formula/timescaledb@2.7.2.rb
@@ -1,4 +1,4 @@
-require_relative 'timescale_base'
+require_relative "../timescaledb_base"
 
 class TimescaledbAT272 < TimescaledbBase
   sha256 "05560e0c37f702a377957adf96d8893c1ff437a8b4fd1478f351e40ba9730897"
