+++ Formula/timescaledb@2.8.0.rb
@@ -1,4 +1,4 @@
-require_relative 'timescaledb_base'
+require_relative "../timescaledb_base"
 
 class TimescaledbAT280 < TimescaledbBase
   sha256 "a3403447805bb97f8b6f38dfdc9dde24c24e4389cf7f2348358774605ad2687e"
