+++ Formula/timescaledb@2.8.1.rb
@@ -1,4 +1,4 @@
-require_relative 'timescaledb_base'
+require_relative "../timescaledb_base"
 
 class TimescaledbAT281 < TimescaledbBase
   sha256 "22a057c4472d23bf08778932e391f38f350ef0307cf99fb8e279c8245667d3e9"
